WIP
Gotta put this off it is getting too crazy
